% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_round.R
\name{adapt_round}
\alias{adapt_round}
\title{round numeric values for presentation in a table}
\usage{
adapt_round(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
round numeric values for presentation in a table
}
