% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable_footers.R
\name{add_ft_footers}
\alias{add_ft_footers}
\title{add footers to a flextable tibbleOne object}
\usage{
add_ft_footers(ft_object, num_headers, table_value_description,
  footnote_markers, footnote_labels, table_notes, tabble_abbrs,
  withhold_footers = FALSE)
}
\description{
add footers to a flextable tibbleOne object
}
