% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable_header.R
\name{add_ft_header}
\alias{add_ft_header}
\title{add headers to flextable object using a kable header vector.}
\usage{
add_ft_header(object, object_data, kable_header = NULL)
}
\arguments{
\item{object}{a flextable object}

\item{object_data}{the dataset used to create `object`}

\item{kable_header}{a header vector for kable objects.}
}
\description{
add headers to flextable object using a kable header vector.
}
