% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_header.R
\name{add_kable_header}
\alias{add_kable_header}
\title{add headers to flextable object using a kable header vector.}
\usage{
add_kable_header(object, escape, bold, header = NULL)
}
\arguments{
\item{object}{a kable object}

\item{escape}{T/F should special characters be ignored?}

\item{bold}{T/F should the header be bold?}

\item{header}{a vector header compatible with kableExtra \code{add_header_above}}
}
\description{
add headers to flextable object using a kable header vector.
}
