% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_kable_headers.R
\name{add_pval_to_header}
\alias{add_pval_to_header}
\title{add a p-value slot into the header for a kable object}
\usage{
add_pval_to_header(x)
}
\arguments{
\item{x}{a header vector for kable}
}
\description{
add a p-value slot into the header for a kable object
}
