% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{capitalize}
\alias{capitalize}
\title{capitalize the first letter of a string (copied from Hmisc)}
\usage{
capitalize(string)
}
\arguments{
\item{string}{String to be capitalized}
}
\description{
capitalize the first letter of a string (copied from Hmisc)
}
