% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_tibble_one_input}
\alias{check_tibble_one_input}
\title{check to make sure that input variables are correctly structured}
\usage{
check_tibble_one_input(object)
}
\arguments{
\item{object}{a tibble_one object.}
}
\description{
check to make sure that input variables are correctly structured
}
