% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tbl_value.R
\name{cmp_pval_noparm}
\alias{cmp_pval_noparm}
\title{compute p-values from non-parametric tests}
\usage{
cmp_pval_noparm(data, variable, ngrps)
}
\description{
compute p-values from non-parametric tests
}
