% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctns_tbl_value.R
\name{ctns_tbl_value}
\alias{ctns_tbl_value}
\title{generate table values for a continuous variable}
\usage{
ctns_tbl_value(variable, data, compute_strat, include.pval,
  include.missinf)
}
\arguments{
\item{variable}{character value, column name of the variable}

\item{data}{a data frame}

\item{compute_strat}{T/F, should table values be stratified?}

\item{include.pval}{T/F, should the table include a column for p-values?}

\item{include.missinf}{T/F, should the table include information on percent of missing values?}
}
\description{
generate table values for a continuous variable
}
