% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_indent_rows}
\alias{find_indent_rows}
\title{find indented rows}
\usage{
find_indent_rows(x)
}
\arguments{
\item{x}{the variable column in a tibble_one kable data object}
}
\description{
find indented rows
}
