% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tbl_value.R
\name{gen_tbl_value}
\alias{gen_tbl_value}
\title{generate table one values for a column variable}
\usage{
gen_tbl_value(variable, type, data, include.pval, include.freq,
  include.allcats)
}
\arguments{
\item{variable}{character value, column name of the variable}

\item{type}{character value, should be numeric, integer, or factor}

\item{data}{a data frame}

\item{include.pval}{T/F, should the table include a column for p-values?}

\item{include.freq}{T/F, should frequency values be included for categorical variables?}

\item{include.allcats}{T/F, should all categories be included for categorical variables?}
}
\description{
generate table one values for a column variable
}
