% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yyy_funs.R
\name{get_abbrs}
\alias{get_abbrs}
\title{If abbreviations for this column are null, return missing value.
Otherwise, return the abbreviations for the column}
\usage{
get_abbrs(data, col)
}
\arguments{
\item{data}{a dataframe.}

\item{col}{a column name in `data`}
}
\description{
If abbreviations for this column are null, return missing value.
Otherwise, return the abbreviations for the column
}
