% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yyy_funs.R
\name{get_groups}
\alias{get_groups}
\title{If group of this column is null, return missing value.
Otherwise, return the group of the column}
\usage{
get_groups(data, col)
}
\arguments{
\item{data}{a dataframe.}

\item{col}{a column name in `data`}
}
\description{
If group of this column is null, return missing value.
Otherwise, return the group of the column
}
