% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_label}
\alias{get_label}
\title{If var_label of this column is null, return col.
Otherwise, return the var_label of the column}
\usage{
get_label(data, col)
}
\arguments{
\item{data}{a dataframe.}

\item{col}{a column name in \code{data}}
}
\description{
If var_label of this column is null, return col.
Otherwise, return the var_label of the column
}
