% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yyy_funs.R
\name{get_notes}
\alias{get_notes}
\title{If note of this column is null, return missing value.
Otherwise, return the notes for the column}
\usage{
get_notes(data, col)
}
\arguments{
\item{data}{a dataframe.}

\item{col}{a column name in `data`}
}
\description{
If note of this column is null, return missing value.
Otherwise, return the notes for the column
}
