% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yyy_funs.R
\name{get_units}
\alias{get_units}
\title{If unit of this column is null, return missing value.
Otherwise, return the unit of the column}
\usage{
get_units(data, col)
}
\arguments{
\item{data}{a dataframe.}

\item{col}{a column name in `data`}
}
\description{
If unit of this column is null, return missing value.
Otherwise, return the unit of the column
}
