% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hline_header.R
\name{hline_header}
\alias{hline_header}
\title{Creat line for merged cells in the header of the table}
\usage{
hline_header(x, border = NULL, bottom = F)
}
\arguments{
\item{x}{a flextable object}

\item{border}{border defined by a call to fp_border}

\item{bottom}{T/F, should bottom border of the header be changed?}
}
\description{
Creat line for merged cells in the header of the table
}
