% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kibble_one.R
\name{kibble_one}
\alias{kibble_one}
\title{pass a tibble_one object into kable}
\usage{
kibble_one(object, format = NULL, use.groups = TRUE, ...)
}
\arguments{
\item{object}{a tibble_one object}

\item{format}{character, format for printing}

\item{use.groups}{T/F, should rows be grouped?}

\item{...}{arguments passed to kable function}
}
\description{
pass a tibble_one object into kable
}
