% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{compute mean and standard deviation of continuous variable}
\usage{
mean_sd(variable)
}
\arguments{
\item{variable}{numeric vector}
}
\description{
compute mean and standard deviation of continuous variable
}
