% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tbl_value.R
\name{median_iqr}
\alias{median_iqr}
\title{compute median and specified percentiles of continuous variable}
\usage{
median_iqr(variable)
}
\arguments{
\item{variable}{numeric vector.}
}
\description{
compute median and specified percentiles of continuous variable
}
