% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_pt.R
\name{medn_pt}
\alias{medn_pt}
\title{compute median and specified percentiles of continuous variable}
\usage{
medn_pt(variable, lower = 0.25, upper = 0.75)
}
\arguments{
\item{variable}{numeric vector.}

\item{lower}{lower percentile to display in table.}

\item{upper}{upper percential to display in table.}
}
\description{
compute median and specified percentiles of continuous variable
}
