% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse_tb1_formula}
\alias{parse_tb1_formula}
\title{tibbleOne formula handler}
\usage{
parse_tb1_formula(formula, data)
}
\arguments{
\item{formula}{a one sided \code{formula} object with row variables
to the left and stratification variables to the right of a
| symbol. For example, formula = ~ x + y | z.}

\item{data}{a data frame to evaluate formula terms in.}
}
\description{
\code{\link[=tibble_one]{tibble_one()}} accepts formula input.
This function translates a formula input value into three
types of table one variables (e.g., row, stratification, and by).
It is unlikely that you will want to use this function directly,
but may wish to use it to debug a formula object that is not
providing the tibble one results you were expecting.
}
\examples{

df = data.frame(a = 1, b = 2, c = 3)

get_tb1_vars(~ a + b | c, df)

# identical to

get_tb1_vars(~ . | c, df)

# specify a by variable using *

get_tb1_vars(~ . | b*c, df)

}
