% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbc_meta}
\alias{pbc_meta}
\title{Primary Biliary Cirrhosis meta data}
\format{a list containing a data frame with 9 rows and 8 variables.
\itemize{
\item variable: variable name - this is the column name of the variable.
\item label: variable labels - this is presented in tables
\item type: type of variable (numeric or factor)
\item unit: units for continuous variables
\item group: a group identifier for each variable
\item abbr: abbreviations associated with the label of a variable
\item note: strings that will be place in tables as a footnote
\item labels:
}}
\usage{
pbc_meta
}
\description{
This is a meta data object that pairs with \code{pbc_tbl1}.
}
\keyword{datasets}
