% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tibble_one}
\alias{print.tibble_one}
\title{print a tibble_one object}
\usage{
\method{print}{tibble_one}(x, ...)
}
\arguments{
\item{x}{an object to print}

\item{...}{other arguments passed to other functions}
}
\description{
print a tibble_one object
}
