% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_labelled.R
\name{select_labelled}
\alias{select_labelled}
\title{Select and label variables}
\usage{
select_labelled(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{name-value pairs of variable labels}
}
\description{
Select and label variables
}
