% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variable_abbreviation.R
\name{set_variable_abbrs}
\alias{set_variable_abbrs}
\title{set the abbreviations for a variable}
\usage{
set_variable_abbrs(data, ...)
}
\arguments{
\item{data}{a data frame containing the columns that will be labeled}

\item{...}{a named vector. Names are variables, values are labels. For example, gfr_variable = 'eGFR = estimated glomerular filtration rate' indicates that the gfr_variable has the given abbreviation.}
}
\description{
set the abbreviations for a variable
}
