% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variable_yyy_funs.R
\name{set_variable_groups}
\alias{set_variable_groups}
\title{set variable groups in a data frame}
\usage{
set_variable_groups(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{name-value pairs of variable groups and names (see examples)}
}
\description{
set variable groups in a data frame
}
