% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_funs.R
\name{set_variable_labels}
\alias{set_variable_labels}
\alias{set_variable_groups}
\alias{set_variable_abbrs}
\alias{set_variable_notes}
\alias{set_variable_units}
\title{Set variable attributes}
\usage{
set_variable_labels(data, ...)

set_variable_groups(data, ...)

set_variable_abbrs(data, ...)

set_variable_notes(data, ...)

set_variable_units(data, ...)
}
\arguments{
\item{data}{a data frame or \code{\link[tibble]{tibble}}.}

\item{...}{name-value pairs of variable groups and names (see examples)}
}
\value{
a data frame or \code{\link[tibble]{tibble}}.
}
\description{
These functions allow you to embed attributes in data
 so that you only need to think about them once. \code{\link{tibble_one}}
 will handle attributes of data automatically, e.g., replacing variable
 names with variable labels, and placing variable acronyms
 (i.e., abbreviations) at the bottom of the table in a footnote.
 Use `set_variable_labels()` to set the values that will represent
 variables in table one. Use `set_variable_groups()` to change the
 variables that are listed in the variable categories of table one.
 Use `set_variable_notes` to add descriptions of variables that will
 be placed at the bottom of table one as a footnote.
 Use `set_variable_abbrs` to indicate what acronyms in variable
 labels mean (see examples).
 Use `set_variable_units` to indicate the unit of measurement for
 continuous variables.
}
\details{
For `set_variable_labels`, names are variables and values are labels.
 For example, writing `gfr_variable` = "estimated GFR" as an input to
 `set_variable_labels` will set the label for `gfr_variable` as the
 indicated string. Since GFR is an acronym, we would also want to use
 `set_variable_abbrs()` and say gfr_variable = c("GFR" =
 "glomerular filtration rate").
}
\examples{
library(magrittr)
library(knitr)
library(kableExtra)

df <- data.frame(
  gfr = rnorm(n = 100, mean = 100, sd = 30),
  sbp = rnorm(n = 100, mean = 120, sd = 20),
  grp = rbinom(n = 100, size = 1, prob = 2/5)
) \%>\%
  set_variable_labels(
    gfr = 'Estimated GFR',
    sbp = 'Systolic BP',
    grp = 'Arbitrary grouping variable'
  ) \%>\%
  set_variable_units(
    gfr = 'mL/min/1.73 m2',
    sbp = 'mm Hg'
  ) \%>\%
  set_variable_abbrs(
    gfr = c("GFR" = "glomerular filtration rate", "min" = 'minute'),
    sbp = c("BP" = "blood pressure")
  ) \%>\%
  set_variable_notes(
    sbp = "blood pressure was measured on the left arm by trained personnel"
  )

tibble_one(
  data = df,
  formula =  ~ gfr + sbp | grp
) \%>\%
  to_kable(escape = FALSE) \%>\%
  kable_styling(full_width = FALSE)
}
\concept{set variable attributes}
