% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variable_yyy_funs.R
\name{set_variable_notes}
\alias{set_variable_notes}
\title{set footnotes for a variable}
\usage{
set_variable_notes(data, ...)
}
\arguments{
\item{data}{a data frame containing the columns that will be given notes.}

\item{...}{a named vector. Names are variables, values are labels. For example, bp_variable = 'blood pressure was measured on the left arm' indicates that the bp_variable has the given note.}
}
\description{
set footnotes for a variable
}
