% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variable_yyy_funs.R
\name{set_variable_units}
\alias{set_variable_units}
\title{set the units for a continuous variable}
\usage{
set_variable_units(data, ...)
}
\arguments{
\item{data}{a data frame containing the columns that will be labeled}

\item{...}{a named vector. Names are variables, values are labels. For example, age = '10 years' indicates that the age variable is in 10 year units.}
}
\description{
set the units for a continuous variable
}
