% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_trim.R
\name{str_split_trim}
\alias{str_split_trim}
\title{split, unlist, and trim a string based on a fixed pattern.}
\usage{
str_split_trim(string, pattern)
}
\arguments{
\item{string}{numeric vector}

\item{pattern}{a fixed pattern to split by}
}
\description{
split, unlist, and trim a string based on a fixed pattern.
}
