% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_word.R
\name{to_word}
\alias{to_word}
\title{pass a tibble_one object into flextable}
\usage{
to_word(object, use_groups = TRUE, font_size = 12,
  apply_format_steps = TRUE, include_1st_header = TRUE,
  include_2nd_header = TRUE, include_3rd_header = TRUE, ...)
}
\arguments{
\item{object}{a tibble_one object}

\item{use_groups}{T/F, should rows be grouped?}

\item{font_size}{the size of font in the table.}

\item{apply_format_steps}{T/F, whether formatting steps should be applied to make the table look nice.}

\item{include_1st_header}{T/F, should bottom header be included?}

\item{include_2nd_header}{T/F, should middle header be included?}

\item{include_3rd_header}{T/F, should top header be included?}

\item{...}{arguments passed to flextable function}
}
\description{
pass a tibble_one object into flextable
}
