% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variable_groups.R
\name{var_group}
\alias{var_group}
\title{get the group attribute of a variable}
\usage{
var_group(x)
}
\arguments{
\item{x}{an object with a group attribute value.}
}
\description{
get the group attribute of a variable
}
