% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vibble}
\alias{vibble}
\title{named vector to tibble}
\usage{
vibble(vector)
}
\arguments{
\item{vector}{a named vector}
}
\description{
turn a named vector into a wide tibble.
This is similar to and connected with \code{tibble::enframe}.
However, the \code{vibble} function will make a wider versus
a longer tibble.
}
