% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_round.R
\name{adapt_round}
\alias{adapt_round}
\title{Adaptive rounding for tables}
\usage{
adapt_round(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a character vector comprising rounded values.
}
\description{
Adaptive rounding for tables
}
\examples{
adapt_round(c(0.12, 10.12, 100.12))
}
