% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta.R
\name{build_meta}
\alias{build_meta}
\title{creates a dataset that describes the characteristics of another dataset}
\usage{
build_meta(data, expand_binary_catgs = FALSE)
}
\arguments{
\item{data}{a data frame with any combination of the following attributes: `label`, `unit`, `group`, `abbrs`, and `notes`. Columns in the meta data are based on these attributes.}

\item{expand_binary_catgs}{T/F, should all categories be included for binary categorical variables? (This only applies to binary variables.)}
}
\description{
creates a dataset that describes the characteristics of another dataset
}
