% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tbl_value.R
\name{catg_tbl_value}
\alias{catg_tbl_value}
\title{generate table values for a categorical variable}
\usage{
catg_tbl_value(variable, data, stratified_table, include.pval = TRUE,
  include.freq = FALSE, expand_binary_catgs = FALSE,
  include.missinf = FALSE)
}
\arguments{
\item{variable}{character value, column name of the variable}

\item{data}{a data frame}

\item{stratified_table}{T/F, should table values be stratified?}

\item{include.pval}{T/F, should the table include a column for p-values?}

\item{include.freq}{T/F, should frequency values be included for categorical variables?}

\item{expand_binary_catgs}{T/F, should all categories be included for categorical variables?}

\item{include.missinf}{T/F, should the table include information on percent of missing values?}
}
\description{
generate table values for a categorical variable
}
