% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tbl_value.R
\name{ctns_tbl_value}
\alias{ctns_tbl_value}
\title{generate mean and standard deviation values for a continuous variable}
\usage{
ctns_tbl_value(data, variable, ctns_fun, pval_fun, stratified_table,
  include_pval)
}
\arguments{
\item{data}{a data frame}

\item{variable}{character value, column name of the variable}

\item{stratified_table}{T/F, should table values be stratified?}

\item{include_pval}{T/F, should the table include a column for p-values?}

\item{include.missinf}{T/F, should the table include information on percent of missing values?}
}
\description{
generate mean and standard deviation values for a continuous variable
}
