% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tbl_value.R
\name{gen_tbl_value}
\alias{gen_tbl_value}
\title{generate table one values for a column variable}
\usage{
gen_tbl_value(data, variable, var_type, fun_type, test_type,
  stratified_table, include_pval, include_freq, expand_binary_catgs)
}
\arguments{
\item{data}{a data frame}

\item{variable}{character value, column name of the variable}

\item{include_pval}{T/F, should the table include a column for p-values? If p-values are included, factor variables are handled using chi-square tests, continuous variables are handled using t-tests or ANOVA, depending on the number of categories in the table stratification.}

\item{include_freq}{T/F, should frequency values be included for categorical variables?}

\item{expand_binary_catgs}{T/F, should all categories be included for binary categorical variables? (This only applies to binary variables.)}

\item{type}{character value, should be numeric, integer, or factor}
}
\description{
generate table one values for a column variable
}
