% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_flex_footers.R
\name{parse_flex_footers}
\alias{parse_flex_footers}
\title{determine which type of symbols to use for flextable footnotes}
\usage{
parse_flex_footers(footnote_notation)
}
\arguments{
\item{footnote_notation}{character value indicating footnote symbols to use in tables. Eligible values are `symbol`, `number`, and `alphabet`.}
}
\description{
determine which type of symbols to use for flextable footnotes
}
