% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_specs.R
\name{parse_specs}
\alias{parse_specs}
\title{handle specifications of table values and table tests}
\usage{
parse_specs(specs_default, specs_analyst)
}
\arguments{
\item{specs_default}{named vector of character values. Names should be variables, while values should all be `default`. Only variables that are in the user-specified formula for `tibble_one` should be included in the `specs_default` vector.}

\item{specs_analyst}{named vector of character values. Names should be variables, while values should be specs (e.g., `mean`, `median` for values, and `params` or `noparm` for tests).}

\item{type}{character value, should be numeric, integer, or factor}
}
\description{
handle specifications of table values and table tests
}
