% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_one.R
\name{tibble_one}
\alias{tibble_one}
\title{make a tidy data frame with table one information}
\usage{
tibble_one(data, strat = NULL, row.vars = NULL, include.pval = TRUE,
  include.freq = FALSE, include.allcats = FALSE,
  include.missinf = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{strat}{a character vector indicating the column name in data that will be used to stratify the table}

\item{row.vars}{a character vector indicating column names of row variables in the table. If unspecified, all columns are used.}

\item{include.pval}{T/F, should the table include a column for p-values?}

\item{include.freq}{T/F, should frequency values be included for categorical variables?}

\item{include.allcats}{T/F, should all categories be included for categorical variables?}

\item{include.missinf}{T/F, should the table include information on percent of missing values?}
}
\description{
make a tidy data frame with table one information
}
