% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleOne.R
\docType{package}
\name{tibbleOne-package}
\alias{tibbleOne}
\alias{tibbleOne-package}
\title{tibbleOne: Because nobody likes making Table 1.}
\description{
To learn more about tibbleOne, start with the vignettes:
\code{browseVignettes(package = "tibbleOne")}
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (0000-0001-7399-2299)

Authors:
\itemize{
  \item Boyi Guo
}

}
