% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleOne.R
\docType{package}
\name{tibbleOne-package}
\alias{tibbleOne}
\alias{tibbleOne-package}
\title{tibbleOne: tidy characteristics tables}
\description{
In many academic papers, table 1 shows participant
characteristics, often stratified by a categorical variable
such as treatment group. There are many excellent packages
available to compute the numbers in table 1. This package
focuses on getting those numbers into a nice format that
works with R Markdown. Additionally, there is a fair amount
of diversity in researchers' preferred writing mediums.
Some may prefer LaTeX, while others want to work in Microsoft
Word. Recently, html documents have grown more common
for research papers. tibbleOne is meant to be applicable for
each of these settings, and should meet the needs of most
studies.

To learn more about tibbleOne, start with the vignettes:
\code{browseVignettes(package = "tibbleOne")}
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (0000-0001-7399-2299)

Other contributors:
\itemize{
  \item Boyi Guo \email{boyiguo1@uab.edu} [contributor]
}

}
