% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_kable.R
\name{to_kable}
\alias{to_kable}
\title{pass a tibble_one object into kable}
\usage{
to_kable(object, format = NULL, use.groups = TRUE,
  include_1st_header = TRUE, include_2nd_header = TRUE,
  include_3rd_header = TRUE, ...)
}
\arguments{
\item{object}{a tibble_one object}

\item{format}{character, format for printing}

\item{use.groups}{T/F, should rows be grouped?}

\item{include_1st_header}{T/F, should bottom header be included?}

\item{include_2nd_header}{T/F, should middle header be included?}

\item{include_3rd_header}{T/F, should top header be included?}

\item{...}{arguments passed to kable function}
}
\description{
pass a tibble_one object into kable
}
