% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_word.R
\name{to_word}
\alias{to_word}
\title{pass tibble_one to flextable}
\usage{
to_word(object, font_size = 11, use_groups = TRUE,
  indent_groups = TRUE, footnote_notation = "symbol",
  include_1st_header = TRUE, include_2nd_header = TRUE,
  include_3rd_header = TRUE, ...)
}
\arguments{
\item{object}{a tibble_one object}

\item{font_size}{the size of font in the table.}

\item{use_groups}{T/F, should rows be grouped?}

\item{indent_groups}{T/F, should entries within groups be indented? (this has no effect if \code{use_groups} is \code{FALSE})}

\item{footnote_notation}{character value indicating footnote symbols to use in tables. Eligible values are \code{symbol}, \code{number}, and \code{alphabet}.}

\item{include_1st_header}{T/F, should bottom header be included?}

\item{include_2nd_header}{T/F, should middle header be included?}

\item{include_3rd_header}{T/F, should top header be included?}

\item{...}{arguments passed to flextable function}
}
\description{
pass tibble_one to flextable
}
