% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble.R
\name{vibble}
\alias{vibble}
\title{turn a named vector into a wide tibble. This is similar to and connected with `tibble::enframe`. However, the `vibble` function will make a wider versus a longer tibble.}
\usage{
vibble(vector)
}
\arguments{
\item{vector}{a named vector}
}
\description{
turn a named vector into a wide tibble. This is similar to and connected with `tibble::enframe`. However, the `vibble` function will make a wider versus a longer tibble.
}
